requirements.txt
Django==5.2
djangorestframework==3.16.0
django-filter==25.1
django-database-url==1.0.3
whitenoise==6.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.9